Updated openstack/openstack Project: openstack-dev/pbr  9858f5863f9328d4ea5b0fa6055ca796c5aa2626
Fix argument order for assertEqual to (expected, observed)

assertEqual expects that the arguments provided to it should be
(expected, observed). If a particluar order is kept as a convention,
then it helps to provide a cleaner message to the developer if Unit
Tests fail. The following patch fixes this issue.

TrivialFix
Closes-Bug: #1259292

Change-Id: I03c414a763b2d51210aaec362405912f29850e36