Updated openstack/openstack Project: openstack/neutron  beee72044f6751fd47a86116face039287532247
Add an independent function to map segment to hosts

_add_segment_host_mapping_for_segment is to update SegmentHostMapping
when a new segment is created. Currently, this function is for agent.
For agentless topology, there is no agent to get the collection of
hosts. There are other ways to get the collection of hosts, which
should be mapped to the newly created segment.
This patch refactors fucntion _add_segment_host_mapping_for_segment,
and adds an independent function map_segment_to_hosts. So that the
agentless topology can use this function to update SegmentHostMapping
when a new segment is created.

Change-Id: If3cc33acc6a6395d0e67c8a74a770fc93e3702e5
Partially-Implements: blueprint routed-networks