Updated openstack/openstack Project: openstack-infra/project-config  bbbe55379318c1cf31ac25e97488a17399c37c60
Implement optional scenario setting for Ansible jobs

In order to allow the same ansible role jobs to be used for multiple
scenarios, this patch implements an optional job setting which gets
used to modify the tox environment used.

The intent is that the tox configuration used in the target repository
can be specified in the job in order to implement alternative
scenarios.

Change-Id: I2e9c49b6d8181feb24c963d88b2f08e46e8a18c4