Updated openstack/openstack Project: openstack/keystone  7923a4669278c1386bea2520adf097760a3ec4fd
Move logic for catalog driver differences to manager

The templated catalog driver does not support some of the
association cleanup methods. Since these are called in normal
operation, we have to mask the raising of NotImplemented. This
masking is currently done in the controller, but should really
be done in the manager layer.

Testing is added for the methods affected at the manager layer,
and an existig v3 API test is tidied up.

Change-Id: Ie68774fef07678b67eb70b90a0e4b81fc0df88b9