Updated openstack/openstack Project: openstack/tempest  e6915250fe88a01d75223af59aef292853f914d1
Add __pycache__ to .gitignore

This commit adds __pycache__ to .gitignore. __pycache__ directories are
made by python 3 execution and we check un-ignored files after the tests
in the gate jobs. And a lot of __pycache__ directories are listed in the
logs. However, it is expected directories so it should be ignored.

Change-Id: I4cfd3475dd2fe1e5b586a28296983a3914236d70