Updated openstack/openstack Project: openstack/cinder-specs  de9ce578602869522e57dc586812c17612f92886
Migrate volume between backends in an async way

Currently if we migrate an available volume from backend A to
backend B who doesn't refer to the same array as backend A,
we can't attach this volume to a server for real use until
the whole migration is totally completed.
This spec is trying to find a way to migrate volume
asynchronously to make the volume usable immediately after
we issued migration task.

Change-Id: I6410c944ff331fa977527b2265d45ce25a36b3b1
Implements: blueprint async-volume-migration-between-backends