Updated openstack/openstack Project: openstack/neutron-lbaas  54941c8124765915aec5e99579966948eaf4762c
Driver exceptions overriden by plugin

When a driver exception is thrown, the LBaaSv2 plugin catches the
error and throws a generic "An error happened in the driver" message.
The original exception error exists in the neutron service log file
 which isn't accessible to the user and requires admin involvement.
The original exception should be raised by the plugin to make the error
troubleshooting easier.

Change-Id: I0163266fd89fe53e021c6d04d06cd28610527b7e
Closes-Bug: #1596472