Updated openstack/openstack Project: openstack/neutron  e3a221df1e4c1e4795591746941eab7abeb3bb68
Ensure deferred IP fails when host is provided no IP allocated

This formalizes that an IP will be allocated when host binding is
provided on port update or IpAddressGenerationFailure is generated and
the port update fails.  This was always the intent but no test was
added to cover the failure case.

Change-Id: Ice0cf2f88eefb15700b1385817e09db489a3bac4
Partially-Implements: blueprint routed-networks