Updated openstack/openstack Project: openstack/diskimage-builder  b87950742089c8f9ebdcf7bd693621d1dbc2359c
Don't create an ironic-agent image just to delete it

The ironic-agent element doesn't care about the final qcow2/raw/
whatever image the disk-image-create command normally creates, so
previously it was deleting it at the end of the process.  This is
a pretty significant waste of time when building those images, and
instead we can just skip creating the image when building
ironic-agent.

Change-Id: If48f575e795a823c777891f193ebf8bd943aa296