Updated openstack/openstack Project: openstack/ironic  e49cd1af0a0a82efeb8a971645eccc7cc39b26d3
Add support for building ISO for deploy ramdisk

This commit adds support for building and uploading
an ISO image for deploy ramdisk.  This is used by
some drivers in Ironic today. A new variable
IRONIC_DEPLOY_ISO_REQUIRED=true can be set in localrc
to create and upload ISO image for the deploy ramdisk.

This patch is required for iLO ThirdParty CI.

Closes-Bug: 1510914
Change-Id: I03a31490323b4bff9700146844a7009212840c60