Updated openstack/openstack Project: openstack/nova  67fabf5ff6919613f78a24e65d7bd4c8fa95e6f6
XenAPI: UT: Always mock logging configuration

The call to configure logging sets up logging to syslog.  Some
environments, such as the Ubuntu package builder environment,
do not permit this operation.

Logging is made to syslog as the plugin runs in dom0 on XenServer,
not with the rest of Nova, so the oslo logging capabilities can't
be used.

Mock out the call to configure logging to prevent attempts to use
syslog

Change-Id: Ie32691d980aff052498c6c5822828ba9703d412b
Closes-Bug: 1597433