Updated openstack/openstack Project: openstack/diskimage-builder  8fdbbc3413fb6a8958c8b283108ec4b85e0f218c
Check sudoers file after editing

"visudo -c" should be run after the sudoers file has been edited. This
will ensure that the file is still syntactically correct, and exit 1 if
it isn't.  Otherwise, obscure errors can occur later on, and it is
difficult to track them back to this script as the source of the error.

Change-Id: Id0e5114d72c0779952a0c2c2c06696929c6c8b17