Updated openstack/openstack Project: openstack/neutron  b6577fa71dde064e6331b1c8f9d712c9169d9dc1
ML2: don't use IntegrityError for duplicate detection

oslo.db will raise a DBDuplicateEntry if an inner IntegrityError
exception is received for a DB opertaion. This is raised if and only
if the error code is 1062.

This patch validates that we have a DBDuplicateEntry instead of just
catching a very generic IntegrityError exception, which may point to
other issues.

The original commit where this was added was commit 54d79acaab50993f08b50f21019f82f17d38caa3.
This was prior to the support added in oslo.db.

Change-Id: I2ad15ad14afc36659bdc3cb0432aa39ec9eeaa30