Updated openstack/openstack Project: openstack/nova  d99aab56364939281a93685c64c50c1c4abba5b0
Fix stale file handle error in resource tracker

This error can occur when resizing vm and
resource tracker are executed at the same time.

The cause of this bug is that
the file path of a VM disk is changed
temporarily when resizing.

So we have to take into account
the OSError(errno.ESTALE)
in order to avoid above error.

Closes-Bug: #1565399
Change-Id: Idf4af5e7ab82c2e4a08180ba5ce8d996db7c270f