Updated openstack/openstack Project: openstack/keystone  e788bed7a3181b9d63197ac7a412d9ba234a537c
Validate SAML keyfile & certfile options

These options are not allowed to contain commas because we have to pass
them to the xmlsec1 binary as a comma-delimited pair. Despite the note
in the help text, there was no validation that the configuration was
usable before we tried to execute the resulting xmlsec1 command.

Change-Id: I88eba566bbe19fa7121e77ffd3f92e4f3d15bbcf