Updated openstack/openstack Project: openstack/cinder  42f0af58afeade0825db09c9284e49a81692c4a7
Replace int with float to bring math.ceil into effect

The function manage_existing_get_size casts the volume size as an int,
which results in returning the floor of the volume size instead of the
ceiling in PY2. It should cast it as a float, which will correctly result
in the ceiling.

Change-Id: I06c71b8f0aaf4473599207ae171356220c8ddf62
Closes-Bug: #1597705