Updated openstack/openstack Project: openstack/nova  f6f4003dfdac83d1decad56cdf39258e9dc75ec0
neutron: delete VIFs when deallocating networking

In e2eb6659a426bae05a54b93ebf31ff3247125258 we started creating
VirtualInterface objects for instances/ports when allocating
networking in the neutronv2 API code. This is used to support
virtual device tagging.

We need to delete the VIFs when deallocating the networking for
the instance though otherwise we can hit unique constraint failures
when trying to re-use an existing port from one instance to another
since the VIF has a unique constraint on the address which is the
port's MAC and uuid when using neutron.

Change-Id: I2254bad0df3ccc00cd5c9438fa2684e705442e2d
Closes-Bug: #1602357