Updated openstack/openstack Project: openstack/keystone  57a5757482619ed09e9b56dbc4a8011d71001694
Disable warnerrors in setup.cfg temporarily

PBR's warnerrors is supposed to fail a build job when warnings are
present in the documentation. Unfortunately that hasn't been working for
a while.

With an upcoming release of PBR this flag will be fixed and any
warnings will fail docs jobs again. There is therefore a coordinated
effort to disable this flag until the PBR release has been completed and
requirements updated. After which time we can reenable the flag and
ensure docs build successfully.

http://lists.openstack.org/pipermail/openstack-dev/2016-June/097849.html

Change-Id: I0043518bdd09d2b2db5be48a942ae91ffc74a524