Updated openstack/openstack Project: openstack/keystone  5c6b2946de116fd61b54d9e967239e9610075c29
Clean up token binding validation code

This patch makes two changes to the token binding validation code easier
to read and provide a better user experience.

Firstly, "if a != b" is used instead of "if not (a == b)" which is
easier to read.

Secondly, validation failures are included in 401 unauthorized responses
instead of the default 401 message. Because the Unauthorized class is
also a SecurityError, insecure_debug will need to be enabled in
keystone.conf in order to expose these details to the API. So, the user
experience for operators trying to debug their token binding
configuration will be improved, but security is not unnecessarily
weakened.

Change-Id: Icc78cacd39a31a33680f891cde1acf4ff41f6ae7