Updated openstack/openstack Project: openstack/cinder-specs  1792dcef94c2e309552714ff5c5a9d2b439e0e7a
Differentiate thick and thin provisioning

If a driver reports thin_provisioning_support to True and
thick_provisioning_support to True on a pool that can support
both thin and thick volumes, and the user wants to create a
thick volume, the current logic in the scheduler would use
the wrong logic which makes decisions based on
provisioned_capacity_gb and max_over_subscription_ratio.

This spec proposes a solution to address the problem.

Change-Id: I9d78ee92fac962f69b56a7bc149ee113c1b18b3f
Implements: blueprint differentiate-thick-thin-in-scheduler