Updated openstack/openstack Project: openstack/tempest  9d8a8c84364b395645b5ad977d183e7aa622527d
Add sync to avoid the loss of pub key data

On the hard rebooting test, the pub key data can be loss due to the
immediate rebooting. In addition, the existing scenario test manager
is using 'sync' command to avoid the other data loss.
So this patch adds a 'sync' command call.

Change-Id: Ie55ad394cdbbde100eee45026c1e52a6f2825b77
Closes-Bug: #1583987