Updated openstack/openstack Project: openstack-infra/nodepool  e173d5eb81ec61c8fe0bf0f1ec3ac7e9c4a9b7e1
Handle NO_AUTO_CREATE_USER default mode

In 5.7 and later versions of mysql connection mode is defaulted to include
NO_AUTO_CREATE_USER. This means that users without passwords being created
with the grant syntax of "GRANT all ON <db>.* to '<user>' ..." will fail.

Long term moving to the use of "CREATE USER" should happen, but for complete
compatibility, it is better to continue with the noraml "GRANT" syntax as
long as a password is provided.

This patch also cleans up users via the "DROP USER" command instead of just
cleaning up databases (in the case the DB instance is re-used, locally for
example).

Change-Id: I65f28872931ade6c223177008cda9deefaa7a8e3