Updated openstack/openstack Project: openstack/keystone  b103c821e0a6d8b6fba8bc52c14a4d10dc02cb5f
Remove a validate_token_bind call

assert_admin is intended to check the roles of the current token to
ensure that there is the admin role. It is basically the v2 equivalent
of @protected for most operations. o

Because keystone requires the auth_context middleware be in the pipeline
we know for any authenticated call a request has passed through this
middleware and so the token bind has already been validated.

We can therefore remove this secondary call. We probably don't need to
worry about the assert_authenticated call however it will produce a
nicer error message if the middleware is not in place.

Change-Id: I2808f0ca9793e892026f083122d4f7266593c496