Updated openstack/openstack Project: openstack/keystone  c104f9fbc21eccb7e817de4a54b4c2fc77ee35bd
Improve user experience involving token flush

Currently with the use of memcache it is no longer necessary to use
the token_flush command. Running this command with KVS driver enabled
fails and throws a Traceback and NotImplemented errors. For a better
UX, we allow the implementation to pass and log a warning message

Change-Id: I95addc8df3a39135fb3fe3c63b6b21c1c279ace8
Closes-Bug: #1520321