Updated openstack/openstack Project: openstack-infra/nodepool  e345107476a82ddad82ea99e184398e0d0e7e85a
Add some tracking for image sizes

Use dib's DIB_SHOW_IMAGE_USAGE flag to make sure we have a report in
the build logs of files and directories >10MiB in the build image.

Add a new gauge stat to measure the size of the built images.

With this, we can make a dashboard to monitor the trend, and if we see
a spike, we should be able to quickly pull out the reason for it from
the logs.

Change-Id: Id94c7f1fabe2e79b162025c8fc4e5ee243b0a8b9
Depends-On: I255800790a62fed1c82fcd311f1cc29c9867766d