Updated openstack/openstack Project: openstack-dev/devstack  72ed1a171ca94e0076ce99e3af63b14d88349833
Restoring xtrace state for os-brick plugin

The change Ibb7423d243d57852dada0b6298463bbdfc6dc63c that introduced the
os-brick plugin introduced a flaw where the xtrace state wasn't restored
after the end of the plugin's execution. The end behavior is that devstack's
logs were with way less information, difficulting the debugging of the build.
This patch fixes the variable that was intended to hold the xtrace state (it
was using cinder's) and restoring the state at the end of the script.

Change-Id: I47c6c794a9704049b089142eca5603d1183f8a10