Updated openstack/openstack Project: openstack-infra/storyboard-webclient  1df407e2476adc68aaf94fc999ee619d46ca9c17
Allow comments to be edited by their owner

This commit adds a way for the author of a comment to edit it in
the web ui. This behaviour is configurable, and is off by default.

It also adds a way to see the full history of comments which have
been edited (clicking on the "edited on" date will open a modal
showing every previous version of the comment). This is available
to everyone, to make is slightly easier to follow conversations in
the case of editing.

In order to turn this behaviour on, add

    "enableEditableComments": true

to `config.json`.

Change-Id: Idd50721afdc4a0ed97ddec42f380b822f852e4ee
Depends-on: Iabf598eae3aa35e4e53aadfe514fb2b0da37cefc