Updated openstack/openstack Project: openstack-infra/project-config  1487231bf8f3115c85d8b72f58579d85697781a1
Revert "Disable all gate jobs for fuel-mirror master"

This reverts commit b42f1f3af329c1da1bebaff0709da35c2a6c3210.

It turned out that disabling all gate jobs also disables us to merge any patches in master to the directory which is still needed. So the best way is to leave pep8 jobs for master and have fake tox target for pep8.

Change-Id: Icfe4f78f4bee7afeecbfef07fb5331d95f4e963c