Updated openstack/openstack Project: openstack-infra/project-config  a159b303c33d6438ca18215599053dae46b4bacb
add job to apply tags when based on changes in openstack/releases

Add a new job to run when a patch merges in openstack/releases to apply
the tags described in the patch. For now, only apply those tags to the
openstack/release-test repository so we can test the job. A later patch
will remove that restriction so the job can apply the tags to all
repositories.

Change-Id: I2121781197d3ed824fed91cdebd0ee9ebeb353d2
Signed-off-by: Doug Hellmann <doug@doughellmann.com>