Updated openstack/openstack Project: openstack-infra/project-config  ec24ed561d0dec8ca3419ee198024321f676c398
force a fuller update of the repo being cloned for tagging

zuul-cloner may be using an out of date cache, and may have pointed
originally at a repo on a zuul merger that doesn't have all tags. Force
an update of all remotes and fetch all tags to make sure they are
present before we start trying to add any new ones.

Change-Id: I29cdae3ff8195e4b7e35ac486429394f17919bff
Signed-off-by: Doug Hellmann <doug@doughellmann.com>