Updated openstack/openstack Project: openstack-infra/project-config  c74565b01819d6badfe449cac49a9c6af2db0614
Always run linters for OpenStack-Ansible jobs

Currently any changes in the docs/releasenotes folders do not get
lint checked. While this is generally fine, the sphinx configuration
should be lint checked.

This patch ensures that the 'linters' check runs against any
documentation/releasenotes changes to ensure that this is covered.

Change-Id: I012756f08d52b9757b36912a66154dd42ae0afbe