Updated openstack/openstack Project: openstack/heat  48f2e5846e4525eb864b2e0331f2e15ed32f9416
Scheduler: Allow a per-task customisable grace period

In some cases we may want to have different grace periods for different
tasks when we cancel a DependencyTaskGroup due to an error. This change
allows us to pass a function instead of just a value, so that we can
compute the correct grace period for each task.

Change-Id: I48842cc2a0193b8c3dc799f6e61c9243b0264b49