Updated openstack/openstack Project: openstack/neutron-lbaas  b91fb91d776f7692cdb4ef784d9f03eecd6362d6
Implements ProcessMonitor in the haproxy driver

The ProcessMonitor class will monitor spawned external processes.

This patch enhances the HaproxyNSDriver class (v2) to utilize the
external_process module in order to monitor and respawn the haproxy processes
if and when needed.

With this patch the LBaaS agent (v2) will load external_process related options
in order to take a configured action when haproxy process dies unexpectedly.

Closes-Bug: #1565801

Change-Id: I420ca20b2620487909885e0e9f08dae60ebec2bf