Updated openstack/openstack Project: openstack/oslosphinx  5c765d5561031311ec1a9347ff2faafa1383efff
Fixed AttributeError: 'str' object has no attribute 'decode'

When git cmd fails, raw_version_list is set to empty 'str' and then
we are trying to decode it. This patch moves decoding inside try/catch
block, so we are decoding return value of subprocess call, which is bytes
in Python 3.

Change-Id: Ibd2f556355a6413beb21473ebfca3b836cbc1d73