Updated openstack/openstack Project: openstack-infra/gear  2defbeaa1beb9ff9e0e3e31529f1f761ac4fcae5
Parse admin requests 2880 times faster

Both returning the incomplete data from the isComplete method, and
scanning the entire incomplete string for the terminator each time
are O(n^2).

Also, returning the string is *very* slow, accounting for 99.6% of
the speedup; scanning the string each time accounts for 0.6%.

Both changes together find the terminus of a 481MiB response in
about 1.6 seconds compared to 76 minutes prior to this change.

Change-Id: I3614e8f3ff8ad7c3d9ca6da1b520b89cd9d5d603