Updated openstack/openstack Project: openstack/requirements  d6cd2a5579ee2b2be9ec8b3f54e3850a27fd17d1
Bump stevedore minimum to 1.16.0

The behavior of NamedExtensionManager when no extensions could be found
changed in 1.16.0 from raising a KeyError to logging a warning (with an
optional callback to handle the error). Writing code to handle the old
and new behavior is pretty nasty, so bump the minimum such that we only
need to handle the new behavior.

See also
http://lists.openstack.org/pipermail/openstack-dev/2016-July/099398.html

Change-Id: I8aa1ee52ff7de50488acb86e8920da89ddb05771
Partial-bug: #1603542