Updated openstack/openstack Project: openstack/cinder  438b2552b4b24957f32d52ead8ff4aab8d30e6b3
Allow admin project to operate on all quotas

Currently when operating on quotas in a nested
hierarchical environment, the calling project context
must be a member of the same hierarchy as the target
project in order to be authorized, even if it has the
admin role. This is a change in behavior for users who
are used to being able to view and modify any project
quotas as admin without needing a token scoped to the
target project's hierarchy.

This patch allows projects with the cloud admin role to
be able to operate on all project quotas regardless of
its place in project hierarchy.

Change-Id: Ifc0f1c5d06e7ca3aa7d9cf267a40f2a6b5cfc078
Closes-bug: #1597045