Updated openstack/openstack Project: openstack/cinder  6156a4559187a466b45449a75f5fc0d03dff91c0
Update initialize_connection to use versionedobjects

The following patch updates initialize_connection
API to use volume versionedobjects.  Changes were
made to be backwards compatible with older RPC clients.
It only includes changes to the core cinder code.
Changes in the drivers are left to each driver
maintainer to update.

Note that this patch DOES NOT try to use object dot
notation everywhere, since it would increase the
size of the patch.  Instead, it will be done in
subsequent patches.

Co-Authored-By: Thang Pham <thang.g.pham@gmail.com>
Change-Id: Ib45cf5934b09fb2de9c228160bbad39c8a896e26
Partial-Implements: blueprint cinder-objects