Updated openstack/openstack Project: openstack/keystonemiddleware  f2be5181abf60fb84d7ea0b49eeec110e26a03d9
Use jsonutils instead of ast for loading the service catalog

AST! Never AST!

The service catalog set from auth_token middleware is a json blob not a
python blob and should be decoded as such.

This brings up the problem that the service catalog specified in the
tests is not actually valid JSON. In future I'll attempt to change this
over to using the auth_token fixture instead of a custom environment
dictionary.

Change-Id: Ic9ab68f7d41d19d3595a3ddbbb2e233f57ef52c8