Updated openstack/openstack Project: openstack-infra/storyboard-webclient  51d9951950d301570f04af4ec2d84ba888f42f24
Make the notifications block fixed on large and medium screens

Currently when an error occurs, notifications of the error appear
in the top-right of the page. However, this makes it easy to miss
them if you are scrolled to a point where they don't appear on
screen.

This commit fixes that for situations where there is enough screen
space to deal with notifications being fixed in position by fixing
them in the top-right of the screen just below the navbar, rather
than having them able to scroll out of view.

Change-Id: I33fe322c7c84c9c2fa0bb43be50f36db2f90edef