Updated openstack/openstack Project: openstack/heat-specs  b452a0d8d058b80b6f85bfdc2ee55383f4b154f1
Add map_replace spec

Adds a spec for a proposed new intrinsic function.

This is required by TripleO to enable transforming a mapping of
service names to networks, where we have two maps, one of service
names to network names, and another (per node) of network names
to IP addresses.

Change-Id: Id5edeeffd22365479fca537a6044cee31a8272bb