Updated openstack/openstack Project: openstack-infra/storyboard-webclient  67e7296780b8035d179faf1592153f54bb1b3b2c
Allow comments to be edited by their owner

This commit adds a way for the author of a comment to edit it in
the web ui. This behaviour is configurable, and is off by default.

It also adds a way to see the full history of comments which have
been edited (clicking on the "edited on" date will open a modal
showing every previous version of the comment). This is available
to everyone, to make is slightly easier to follow conversations in
the case of editing.

In order to turn this behaviour on, add

    "enableEditableComments": true

to `config.json`.

Change-Id: I284e994b528c267a596189e3db806cb2d83b4ef1